chore(lsp): cleanup ranking expression parser
